Use uint16_t for opcode and oparg

+++ src/pyjion/binarycomp.cpp
@@ -40,7 +40,7 @@ typedef void(__cdecl* JITSTARTUP)(ICorJitHost*);
 using namespace std;
 extern BaseModule g_module;
 
-void PythonCompiler::emit_binary_object(int opcode) {
+void PythonCompiler::emit_binary_object(uint16_t opcode) {
     switch (opcode) {
         case BINARY_ADD:
             m_il.emit_call(METHOD_ADD_TOKEN); break;
@@ -97,7 +97,7 @@ void PythonCompiler::emit_binary_object(int opcode) {
     }
 }
 
-void PythonCompiler::emit_binary_object(int opcode, AbstractValueWithSources left, AbstractValueWithSources right) {
+void PythonCompiler::emit_binary_object(uint16_t opcode, AbstractValueWithSources left, AbstractValueWithSources right) {
     int nb_slot = -1;
     int sq_slot = -1;
     int fallback_token;
@@ -679,7 +679,7 @@ void PythonCompiler::emit_known_binary_op_power(AbstractValueWithSources &left, 
     }
 }
 
-void PythonCompiler::emit_triple_binary_op(int firstOp, int secondOp) {
+void PythonCompiler::emit_triple_binary_op(uint16_t firstOp, uint16_t secondOp) {
     m_il.ld_i4(firstOp);
     m_il.ld_i4(secondOp);
     m_il.emit_call(METHOD_TRIPLE_BINARY_OP);
